Avance CU Consultar estadísticas, Mejoras Interfaz Usuarios
Modificación de script y JPA para Pagos y Cliente
Mejoras de interfaz en Administración de Usuarios
Mejoras de búsqueda en Administración de Usuarios
Interfaz y progreso de implementación Consultar estadísticas

+++ Proyecto/ared/src/main/java/controladores/PantallaRegistrarPagoMaestroController.java
@@ -182,7 +182,7 @@ public class PantallaRegistrarPagoMaestroController implements Initializable, Co
                         pagoMaestro.setMaestro((Maestro) maestros.get(lstMaestros.getSelectionModel().getSelectedIndex()));
                         pagoMaestro.setFecha(new Date());
                         pagoMaestro.setFechaVencimiento(fechaVencimiento);
-                        pagoMaestro.setMonto(Integer.parseInt(txtMonto.getText()));
+                        pagoMaestro.setMonto(Double.parseDouble(txtMonto.getText()));
                         if (pagoMaestro.registrarPago()) {
                             pnlPrincipal.getChildren().add(crearPantalla("/fxml/PantallaRegistrarPagoMaestro.fxml", this.pnlPrincipal, this.pantallaDividida));
                             pantallaDividida.getChildren().add(pnlPrincipal);
